feat: 링크 관리 컴포넌트 분리 및 추가하기 구현 #37
