+++ src/types/link.ts
@@ -3,6 +3,7 @@ export type LinkType = "instagram" | "facebook" | "threads" | "custom";
 export interface Link {
   id: number;
   userId: number;
+  type: LinkType;
   title: string;
   image: string;
   url: string;
